% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm.R
\name{dfm}
\alias{dfm}
\title{Estimates a dynamic factor model based on Doz, Gianone & Reichlin (2011)}
\usage{
dfm(data, r, p, q = r, max_iter = 100, threshold = 1e-04,
  lower_set = 1e-07, rQ = c("none", "identity"), rC = c("none", "upper"),
  ...)
}
\arguments{
\item{data}{Data matrix with time in rows and variables in columns.}

\item{r}{Number of static factors to estimate.}

\item{p}{Lag order for factors. It is assumed that factors follow VAR(p)
model.}

\item{q}{Number of dynamic factors, must be equal to or less than r. Dynamic
factors refer essentially to the number of principal components relevant
for the state covariance estimation.}

\item{max_iter}{Maximum number of iterations in the EM-algorithm.}

\item{threshold}{Threshold for algorithm convergence}

\item{lower_set}{In order to keep an invertible observation covariance
matrix, diagonal values are constrained to be such that
\code{diag(R)[diag(R) < lower_set] <- lower_set}.}

\item{rQ}{Restrictions on system state covariance. Currently, either no
restrictions can be set or \code{Q} can be fixed as identity.}

\item{rC}{Restrictions on factor loading matrix. Currently, either no
restrictions can be set or upper matrix of \code{C} can be set to 0.}

\item{...}{Further arguments that are currently unused}
}
\value{
3 types of factor estimates, namely principal component estimate, two
  step estimate based on PCA and Kalman filtering and QML estimate based on
  EM-algorithm. PCA estimator is not able to deal with missing data.
}
\description{
Estimates a dynamic factor model based on Doz, Gianone & Reichlin (2011)
}
\examples{
\dontrun{
set.seed(314)
x <- matrix(rnorm(50*10), 50, 10)
W <- as.logical(matrix(rbinom(50*10, 1, 0.1), 50, 10))
x[W] <- NA
dfm(x, 2, 2, 1)
}
}
