% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmMS.R
\name{dfmMS}
\alias{dfmMS}
\title{Dynamic factor model with Markov-switching states}
\usage{
dfmMS(data, nf = 1, ns = 2, x0 = rep(0, nf), P0 = diag(1, nf), init,
  ...)
}
\arguments{
\item{data}{Data matrix}

\item{nf}{Number of factors}

\item{ns}{Number of states}

\item{x0}{Initial value for state vector}

\item{P0}{Initial value for state covariance, i.e. uncertainty of the initial
state value vector}

\item{init}{List with initial values for A, F, R, p. If not supplied,
\code{\link{dfm}} function will be run and the estimates will be used as
initial values.}
}
\description{
This function extends in a sense \code{\link{dfm}} in that it allows
observation or transition matrices to follow a Markov-switching
process, so that they are state-dependent.
}
\details{
Note that this method does not implement Chang-Jin Kim algorithm. It is based
on the same idea as the \code{\link{dfm}} function with the adaptation that
Kalman filter is replaced by Kim filter which is able to deal with multiple
states. Optimization is done over a list of parameters by a simple call to
\code{\link{optim}} instead of an EM-algorithm.

Currently, state equation covariance matrix is restricted to identity.
}
