% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Estep}
\alias{Estep}
\title{Computation of the expectation step in the EM-algorithm.}
\usage{
Estep(y, A, C, Q, R, initx, initV, W)
}
\arguments{
\item{y}{Data matrix}

\item{A}{System state matrix}

\item{C}{Observation matrix}

\item{Q}{System state equation variance}

\item{R}{Observation equation variance}

\item{initx}{Initial value for state variable}

\item{initV}{Initial value for state matrix}

\item{W}{Logical matrix with dim(W) = dim(y)
indicating missing observations}
}
\value{
Sufficient statistics used for M-step
}
\description{
Computation of the expectation step in the EM-algorithm.
}
