% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KalmanFilterCpp2}
\alias{KalmanFilterCpp2}
\title{Implementation of a Kalman filter}
\usage{
KalmanFilterCpp2(y, F, Q, R, A, x0, P0)
}
\arguments{
\item{y}{Data matrix}

\item{F}{Observation matrix}

\item{Q}{State covariance}

\item{R}{Observation covariance}

\item{A}{Transition matrix}

\item{x0}{Initial state vector}

\item{P0}{Initial state covariance}
}
\value{
A list with estimates
}
\description{
Implementation of a Kalman filter
}
