% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KalmanSmootherCpp}
\alias{KalmanSmootherCpp}
\title{Runs a Kalman smoother}
\usage{
KalmanSmootherCpp(A, F, R, xitt, xittm, Ptt1, Pttm1)
}
\arguments{
\item{A}{transition matrix}

\item{F}{observation matrix}

\item{R}{Observation covariance}

\item{xitt}{State estimates}

\item{xittm}{State predicted estimates}

\item{Ptt1}{Variance estimates}

\item{Pttm1}{Predicted variance estimates}
}
\value{
List of smoothed estimates
}
\description{
Runs a Kalman smoother
}
