% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/wrappers.R
\name{KimFilter}
\alias{KimFilter}
\alias{KimFilter}
\title{Implementation of Kim (1994) filter, an extension to Kalman filter
for dynamic linear models with Markov-switching. Documentation
is incomplete, rudimentary and needs to be rechecked!}
\usage{
KimFilter(y, R, Q, F1, A1, x0, P0, p)

KimFilter(y, R, Q, F1, A1, x0, P0, p)
}
\arguments{
\item{y}{Data matrix (Txn)}

\item{R}{Error covariance for measurement equation}

\item{Q}{Error covariance for state equation}

\item{F1}{Array of observation matrices, one matrix per state}

\item{A1}{Array of transition matrices, one matrix per state}

\item{x0}{Initial condition for state vector}

\item{P0}{Initial condition for state variance}

\item{p}{Transition probability matrix}

\item{F}{System matrix for measurement equation}

\item{A}{Transition matrix for state equation}
}
\value{
Filtered states and covariances with associated probability matrices.

This function returns the same output as \code{\link{KimFilterCpp}}.
  However, it converts certain elements of the Kim Filter to standard multi-dimensional
  R arrays and so its outputs can be used as inputs for \code{\link{KimSmoother}}.
}
\description{
Implementation of Kim (1994) filter, an extension to Kalman filter
for dynamic linear models with Markov-switching. Documentation
is incomplete, rudimentary and needs to be rechecked!

Kim filtering
}
\seealso{
\code{\link{KimFilterCpp}}
}
