% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KimFilterCpp}
\alias{KimFilterCpp}
\title{Implementation of Kim filter (1994), an extension to Kalman filter
for dynamic linear models with Markov-switching parameters. Currently,
Markov switching is assumed to happen only in observation and/or
transition matrices.}
\usage{
KimFilterCpp(y, R, Q, F1, A1, x0, P0, p)
}
\arguments{
\item{y}{Data matrix (\code{T x n})}

\item{R}{Observation equation covariance}

\item{Q}{State equation covariance}

\item{F1}{Array of observation matrices, one matrix per state}

\item{A1}{Array of transition matrices, one matrix per state}

\item{x0}{Initial condition for state vector}

\item{P0}{Initial condition for state covariance matrix}

\item{p}{Markov transition probability matrix}
}
\value{
a list with estimates
}
\description{
Implementation of Kim filter (1994), an extension to Kalman filter
for dynamic linear models with Markov-switching parameters. Currently,
Markov switching is assumed to happen only in observation and/or
transition matrices.
}
