% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBBsurvey}
\alias{NBBsurvey}
\title{National Bank of Belgium business and consumer surveys}
\format{A time-stamped dataframe (xts object) with}
\source{
\url{http://stat.nbb.be}
}
\description{
National Bank of Belgium business and consumer surveys
}
\author{
National Bank of Belgium
}
