% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{summary.dfm}
\alias{summary.dfm}
\title{Summary information on dynamic factor model estimation}
\usage{
\method{summary}{dfm}(object, ...)
}
\arguments{
\item{x}{An object of \code{dfm} class}

\item{plot}{If \code{TRUE}, returns a bunch of summary plots}
}
\value{
Prints out a summary information following a dynamic factor
model estimation. Also can return summary plots.
}
\description{
Summary information on dynamic factor model estimation
}
